Update mixed precision settings to use bf16 and add metrics module
- Changed default mixed precision from fp16 to bf16 in multiple scripts and configuration files for improved performance.
- Added a new metrics module to evaluate image quality metrics including FID, KID, CLIP-Score, and more.
- Updated README.md to reflect changes in the project structure and added usage instructions for the new metrics.
- Introduced a benchmark case study script for evaluating image generation at different resolutions.
- Updated .gitignore to include outputs_test directory.

+++ app.py
@@ -235,7 +235,7 @@ def build_demo(ckpt_dir: str = "./models") -> gr.Blocks:
                 with gr.Row():
                     mixed_precision = gr.Dropdown(
                         choices=["fp16", "bf16", "no"],
-                        value="fp16",
+                        value="bf16",
                         label="Precision",
                     )
                     enable_reschedule = gr.Checkbox(
